% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_medHMM2}
\alias{forecast_medHMM2}
\title{Function returning expected observations fiven the forecasted states.}
\usage{
forecast_medHMM2(
  s_data,
  object,
  forecast_steps = 1,
  value_range = NULL,
  burn_in = NULL,
  Mx = NULL,
  shift = 1,
  return_all = FALSE
)
}
\description{
Function returning expected observations fiven the forecasted states.
}
