% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_medHMM3}
\alias{forecast_medHMM3}
\title{Function returning expected observations fiven the forecasted states.}
\usage{
forecast_medHMM3(
  s_data,
  object,
  initial_window = 600,
  value_range = NULL,
  burn_in = NULL,
  Mx = NULL,
  shift = 1,
  return_all = FALSE,
  return_full_sequence = FALSE,
  h_step = 5,
  show_progress = TRUE
)
}
\description{
Function returning expected observations fiven the forecasted states.
}
